+++ src/pages/index.js
@@ -7,6 +7,27 @@ import ContentWrapper from "../components/ContentWrapper"
 import { loadState } from '../localStorage'
 import FirstTimeUser from "../components/FirstTimeOnApp/FirstTimeUser"
 import  "../../firebase"
+
+
+
+function initNetlifyIdentity() {
+  console.log("initNetlifyIdentity called.")
+  const script = document.createElement("script");
+  script.src = "https://identity.netlify.com/v1/netlify-identity-widget.js"
+  script.async = true;
+  document.body.appendChild(script);
+}
+
+export function openNetlifyModal() {
+  const netlifyIdentity = window.netlifyIdentity;
+
+  if(netlifyIdentity)
+    netlifyIdentity.open();
+  else
+    console.log('netlifyIdentity not defined')
+}
+
+
 const setHandle =()=>{
   if(loadState('listOfResonsArray')){
    return(false)
@@ -16,10 +37,10 @@ const setHandle =()=>{
 }
 const IndexPage = () => {
   const [firstTime, setFirstTime] = React.useState(setHandle())
-  // useEffect(() => {
-  //   document.body.addEventListener('click', console.log('klik klik klik XD'));
-  // })
-
+  
+  useEffect(() => {
+    initNetlifyIdentity();
+    console.log("iniet Net... End")})
 
   const setHandleFalse =()=>{
     setFirstTime(false)
@@ -28,7 +49,7 @@ return(
   <Layout>
     <SEO title="Home" />
        {firstTime ? <FirstTimeUser setStart={setHandleFalse} /> : <>
-       <MenuPanel /><ContentWrapper />
+      <MenuPanel /><ContentWrapper />
        </>}
   </Layout>
   )
